MQTT_sensor_rework (#16)
* Disable Access point interface on ESP8266

By default AP on ESP8266 is active but it is not necessary for our sensors.

* Add mac address print for debug

* Remove air quality measurements on BME680 on ESP8266

Removed because it is not very accurate and hard to manage. It will be better to add a new sensor which will for example measure CO2 values in room.

* Rework ESP8266 Wifi connection

Sensor now disconnects after writing data to MQTT. This is a preparation for putting sensor in deep sleep mode where it draws less power. Useful for battery powered sensors.

Sensor now asks for room every time it connects to Wifi. Because it cant watch for new room value while disconnected. Also room value is now changed in MariaDB database and then it is sent to sensor.

* Add deepsleep mode for ESP8266 sensors

Sensors on ESP8266 board now enters deepsleep mode in set intervals (set by variable message_interval_ms). In this mode sensor draws minimal power but it needs to be woken up by command. After set interval passes sensor is restarted using GPIO16 wakeup pin connected to ESP8266 reset pin.